requirements.txt
asgiref==3.7.2
certifi==2023.5.7
cloudinary==1.33.0
Django==4.2.2
djangorestframework==3.14.0
Pillow==10.0.0
psycopg2-binary==2.9.6
python-decouple==3.8
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
